Refactor carousel code to be in `index.html` and add stylistic changes
Put all of the 	`html` for carousel into `index.html`. Changed the
styling so it falls more in line with the theme of the website.
Moved buttons and changed their class name so they could sit
on either side of the carousel. Made the necessary changes in the
`carousel.js` file so this will work.
Refactored `sass` variables into a new file