Tweek webpack to host static images
So I think I solved your problem. The static images weren't being
loaded and bundled with webpack. What I did is require them in the
index.js file and then changed the path so that they are all dumped
in `public/img/[name].[ext]` to make it a bit cleaner. I then changed
the `src` paths to reflect the right path when the project is deployed.
I have it working locally so give it a try and see. Let me know if it
works